backend/requirements.txt
aioredis==1.2.0
asgiref==2.3.2
astroid==2.1.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.2.1
Automat==0.7.0
certifi==2018.11.29
channels==2.1.7
channels-redis==2.3.3
constantly==15.1.0
daphne==2.2.5
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.7
django-cors-headers==2.4.1
django-filter==2.1.0
django-filters==0.2.1
django-heroku==0.3.1
django-picklefield==2.0
djangorestframework==3.9.1
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
msgpack==0.6.1
pipenv==2018.11.26
psycopg2==2.7.7
PyHamcrest==1.9.0
pylint==2.2.2
python-decouple==3.1
pytz==2018.9
redis==3.2.0
six==1.12.0
static3==0.7.0
Twisted==18.9.0
txaio==18.8.1
virtualenv==16.3.0
virtualenv-clone==0.5.1
whitenoise==4.1.2
wrapt==1.11.1
zope.interface==4.6.0
